first shot cleaning up Consumable/Food interfaces
