+++ src/entity/Consumable.php
@@ -21,19 +21,21 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\item;
+namespace pocketmine\entity;
+
+use pocketmine\entity\effect\EffectInstance;
 
 /**
- * Interface implemented by objects that can be consumed by players, giving them food and saturation.
+ * Interface implemented by objects that can be consumed by mobs.
  */
-interface FoodSource extends Consumable{
-
-	public function getFoodRestore() : int;
-
-	public function getSaturationRestore() : float;
+interface Consumable{
+	/**
+	 * @return EffectInstance[]
+	 */
+	public function getAdditionalEffects() : array;
 
 	/**
-	 * Returns whether a Human eating this FoodSource must have a non-full hunger bar.
+	 * Called when this Consumable is consumed by mob, after standard resulting effects have been applied.
 	 */
-	public function requiresHunger() : bool;
+	public function onConsume(Living $consumer) : void;
 }
